+++ art_street/lib/main.dart
@@ -1,4 +1,4 @@
-import 'login.dart';
+import 'pages/login.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(MyApp());
@@ -17,4 +17,3 @@ class MyApp extends StatelessWidget {
     );
   }
 }
-
