+++ art_street/lib/pages/singup.dart
@@ -1,0 +1,9 @@
+import 'dart:html';
+import 'package:flutter/material.dart';
+
+class SignUp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold();
+  }
+}
