--- art_street/lib/singup.dart
@@ -1,11 +1,0 @@
-import 'dart:html';
-import 'package:flutter/material.dart';
-
-class SignUp extends StatelessWidget{
-  @override 
-  Widget build(BuildContext context){
-    return Scaffold(
-
-    );
-  }
-}
